[UPUP-986] правки нейминг
