+++ Sources/MUNKit/MobileApi/MUNKMobileApiTargetType.swift
@@ -1,6 +1,6 @@
 //
 //  MUNKMobileApiTargetType.swift
-//  NetworkService
+//  MUNKit
 //
 //  Created by Natalia Luzyanina on 01.04.2025.
 //
