--- Sources/MUNKit/MoyaProvider+Async.swift
@@ -1,64 +1,0 @@
-//
-//  MoyaProvider.swift
-//  NetworkService
-//
-//  Created by Natalia Luzyanina on 01.04.2025.
-//
-
-import Moya
-
-extension MoyaProvider {
-    func request<T: Decodable & Sendable>(target: Target) async throws -> T {
-        return try await withCheckedThrowingContinuation { continuation in
-            request(target) { [weak self] result in
-                switch result {
-                case .success(let response):
-                    self?.handleRequestSuccess(response: response, continuation: continuation)
-                case .failure(let error):
-                    self?.handleRequestFailure(error: error, continuation: continuation)
-                }
-            }
-        }
-    }
-    
-    func request(target: Target) async throws {
-        return try await withCheckedThrowingContinuation { continuation in
-            request(target) { [weak self] result in
-                switch result {
-                case .success(let response):
-                    self?.handleRequestSuccess(response: response, continuation: continuation)
-                case .failure(let error):
-                    self?.handleRequestFailure(error: error, continuation: continuation)
-                }
-            }
-        }
-    }
-    
-    private func handleRequestSuccess<T: Decodable & Sendable>(response: Response, continuation: CheckedContinuation<T, Error>) {
-        do {
-            let filteredResponse = try response.filterSuccessfulStatusCodes()
-            let decodedResponse = try filteredResponse.map(T.self)
-            
-            continuation.resume(returning: decodedResponse)
-        } catch let error {
-            continuation.resume(throwing: error)
-        }
-    }
-    
-    private func handleRequestSuccess(response: Response, continuation: CheckedContinuation<Void, Error>) {
-        do {
-            _ = try response.filterSuccessfulStatusCodes()
-            continuation.resume()
-        } catch let error {
-            continuation.resume(throwing: error)
-        }
-    }
-    
-    private func handleRequestFailure<T: Decodable>(error: MoyaError, continuation: CheckedContinuation<T, Error>) {
-        continuation.resume(throwing: error)
-    }
-    
-    private func handleRequestFailure(error: MoyaError, continuation: CheckedContinuation<Void, Error>) {
-        continuation.resume(throwing: error)
-    }
-}
