+++ Sources/MUNKit/Plugins/AccessTokenPlugin.swift
@@ -1,6 +1,6 @@
 //
 //  AccessTokenPlugin.swift
-//  NetworkService
+//  MUNKit
 //
 //  Created by Natalia Luzyanina on 01.04.2025.
 //
